+++ longtests/testthat/test_cBioPortalData.R
@@ -12,10 +12,13 @@ test_that("cBioPortal API is working with most studies", {
         complete[[api_stud]] <- is(
             tryCatch({
                 cBioPortalData(
-                    cbioportal, studyId = api_stud, genePanelId = "IMPACT341"
+                    cbio, studyId = api_stud, genePanelId = "IMPACT341"
                 )
             }, error = function(e) conditionMessage(e)),
             "MultiAssayExperiment"
+        )
+        removeDataCache(
+            cbio, studyId = api_stud, genePanelId = "IMPACT341", dry.run = FALSE
         )
     }